// server/application/microservices/assessment/src/common/errors/error-codes.enum.ts
/**
 * Standardized Error Codes for EdForge Academic Module
 * 
 * Format: <MODULE>_<NUMBER>
 * Ranges:
 *   1xxx - Classroom errors
 *   2xxx - Enrollment errors
 *   3xxx - Assignment errors
 *   4xxx - Grading errors
 *   5xxx - Attendance errors
 *   9xxx - General/system errors
 */

export enum AcademicErrorCode {
  // Classroom Errors (1xxx)
  CLASSROOM_NOT_FOUND = 'CLASSROOM_1001',
  CLASSROOM_AT_CAPACITY = 'CLASSROOM_1002',
  CLASSROOM_INVALID_SCHEDULE = 'CLASSROOM_1003',
  CLASSROOM_CODE_DUPLICATE = 'CLASSROOM_1004',
  CLASSROOM_INACTIVE = 'CLASSROOM_1005',
  CLASSROOM_ARCHIVED = 'CLASSROOM_1006',
  
  // Student Enrollment Errors (2xxx)
  STUDENT_ALREADY_ENROLLED = 'ENROLLMENT_2001',
  STUDENT_NOT_ENROLLED = 'ENROLLMENT_2002',
  STUDENT_NOT_FOUND = 'ENROLLMENT_2003',
  ENROLLMENT_CAPACITY_EXCEEDED = 'ENROLLMENT_2004',
  ENROLLMENT_PREREQUISITE_NOT_MET = 'ENROLLMENT_2005',
  
  // Assignment Errors (3xxx)
  ASSIGNMENT_NOT_FOUND = 'ASSIGNMENT_3001',
  ASSIGNMENT_ALREADY_PUBLISHED = 'ASSIGNMENT_3002',
  ASSIGNMENT_INVALID_DATES = 'ASSIGNMENT_3003',
  ASSIGNMENT_EXCEEDS_MAX_SIZE = 'ASSIGNMENT_3004',
  ASSIGNMENT_ARCHIVED = 'ASSIGNMENT_3005',
  ASSIGNMENT_PAST_DUE = 'ASSIGNMENT_3006',
  
  // Grading Errors (4xxx)
  GRADE_NOT_FOUND = 'GRADE_4001',
  GRADE_ALREADY_EXISTS = 'GRADE_4002',
  GRADE_INVALID_SCORE = 'GRADE_4003',
  GRADE_SCALE_NOT_FOUND = 'GRADE_4004',
  GRADE_CONCURRENT_MODIFICATION = 'GRADE_4005',
  GRADE_EXCEEDS_MAX_POINTS = 'GRADE_4006',
  GRADE_ALREADY_PUBLISHED = 'GRADE_4007',
  
  // Attendance Errors (5xxx)
  ATTENDANCE_NOT_FOUND = 'ATTENDANCE_5001',
  ATTENDANCE_ALREADY_RECORDED = 'ATTENDANCE_5002',
  ATTENDANCE_INVALID_DATE = 'ATTENDANCE_5003',
  ATTENDANCE_FUTURE_DATE = 'ATTENDANCE_5004',
  
  // Stream/Post Errors (6xxx)
  POST_NOT_FOUND = 'STREAM_6001',
  POST_UNAUTHORIZED = 'STREAM_6002',
  COMMENT_NOT_FOUND = 'STREAM_6003',
  
  // General Errors (9xxx)
  VALIDATION_FAILED = 'GENERAL_9001',
  CONCURRENT_MODIFICATION = 'GENERAL_9002',
  UNAUTHORIZED_ACCESS = 'GENERAL_9003',
  RESOURCE_NOT_FOUND = 'GENERAL_9004',
  INTERNAL_ERROR = 'GENERAL_9999'
}